dynamic scout
scout now goes to every unknown location on the map as well as the enemy base and keeps cycling through these locations to give us up-to-date info on potential new bases of the enemy.

This avoids the "infinite while loop" situations where the enemy moves their base to a location not in the path to their start location or close to it.